extracted sort and filter from client to hook
